requirements.txt
torch
transformers
bitsandbytes
peft
langchain
chromadb
telebot
googletrans==3.1.0a0
pandas
requests
Pillow